fix: Fixed the issue on response of offer's endpoints.
